Removed the dropdown menu, and also changed the patient demographics tile as requested
