+++ src/components/Header/Header.component.jsx
@@ -1,37 +1,17 @@
 // Import: Dependencies
-import React, { useState } from "react";
-
-// Import: Assets
-import { ReactComponent as DropdownIcon } from "../../assets/img/icons/dropdown.svg";
-import { ReactComponent as UpArrowIcon } from "../../assets/img/icons/uparrow.svg";
+import React from "react";
 
 // Import: Elements
-import { Container, DropdownButton, Logo, Wrapper } from "./Header.elements";
-
-// Import: Components
-import { Dropdown } from "../index";
+import { Container, Logo, Wrapper } from "./Header.elements";
 
 // Component: Header
 export default function Header() {
-  // State = isDropdownOpen
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-
   return (
     <Container>
       <Wrapper>
         <Logo>
-          <DropdownButton
-            onClick={() =>
-              setIsDropdownOpen((isDropdownOpen) => !isDropdownOpen)
-            }
-          >
-            {!isDropdownOpen ? <DropdownIcon /> : <UpArrowIcon />}
-          </DropdownButton>
-
-          <span>Lorenzo Report</span>
+           <span>Lorenzo Report</span>
         </Logo>
-
-        <Dropdown isDropdownOpen={isDropdownOpen} />
       </Wrapper>
     </Container>
   );
