+++ src/pages/subPages/DiagnosisOfDeath/renders/DODAttendance/DODAttendance.component.jsx
@@ -11,7 +11,7 @@ import {
   Heading,
   Render,
   Section,
-} from "./DODSeventh.elements";
+} from "./DODattendance.elements";
 
 // Import: Components
 import {
